fix(proxy): normalize Responses user identifiers
