+++ src/providers/plugins/sso/proxy/plugins/index.ts
@@ -16,7 +16,10 @@ import { RequestSanitizerPlugin } from './request-sanitizer.plugin.js';
 import { ClaudeRequestNormalizerPlugin } from './claude-request-normalizer.plugin.js';
 import { KimiRequestNormalizerPlugin } from './kimi-request-normalizer.plugin.js';
 import { CodexEncryptedContentSanitizerPlugin } from './codex-encrypted-content-sanitizer.plugin.js';
-import { VsCodeRequestNormalizerPlugin } from './vscode-request-normalizer.plugin.js';
+import {
+  ResponsesUserNormalizerPlugin,
+  VsCodeRequestNormalizerPlugin,
+} from './responses-user-normalizer.plugin.js';
 import { LoggingPlugin } from './logging.plugin.js';
 import { SSOSessionSyncPlugin } from './sso.session-sync.plugin.js';
 
@@ -37,7 +40,7 @@ export function registerCorePlugins(): void {
   registry.register(new KimiRequestNormalizerPlugin()); // Priority 14 - caps Kimi output token requests for upstream limits
   registry.register(new RequestSanitizerPlugin()); // Priority 15 - strips unsupported reasoning params
   registry.register(new CodexEncryptedContentSanitizerPlugin()); // Priority 16 - strips deployment-bound encrypted Responses reasoning state
-  registry.register(new VsCodeRequestNormalizerPlugin()); // Priority 17 - constrains VS Code user identifiers
+  registry.register(new ResponsesUserNormalizerPlugin()); // Priority 17 - constrains Responses user identifiers
   registry.register(new HeaderInjectionPlugin());
   registry.register(new LoggingPlugin()); // Always enabled - logs to log files at INFO level
   registry.register(new SSOSessionSyncPlugin()); // Priority 100 - syncs sessions via multiple processors
@@ -58,6 +61,7 @@ export {
   ClaudeRequestNormalizerPlugin,
   KimiRequestNormalizerPlugin,
   CodexEncryptedContentSanitizerPlugin,
+  ResponsesUserNormalizerPlugin,
   VsCodeRequestNormalizerPlugin,
   LoggingPlugin,
 };
